all Files used in v0.3 of Manual-Ranking.bat
This addition contains all the files used in the current version of the Manual Ranker I made.  Of course the brokenSELECT.def and stages will not use the same file path as they do now but that can be changed later.  v0.3 isn't set in stone and neither is the elevator-music.mp3.  There are some obvious areas that can be fixed.  No game interaction yet.