v0.6
A lot of edits have been made.  including edits to the select.def i included for testing.  now it gets the category properly and it shouldn't be long until the authors name can be shoved in there.  I also accidently put a timed reset on my drive mount letters so wish me luck on that.