Version 18/08/20
**I found out I messed up the version numbers so i'm going by number now**

In this update, I have added the ability to select both chars.  These can be random by putting "randomselect" in the player field or by selecting a name for the program to search for.  The program will remember the choice of random or "chosen" and carry on to the next match.

When selecting to use an assigned character the program will get the line for the first occurance of that name and then set it as the line number to be the "targetchar" line.  this makes it so any other information linked to the character can be retreived.

finally added proper credit to Inktrebuchet in the header with link.

There is an exceptional error where if the name of "player2" contains a bracket the program will crash but at the moment this is a 1 in ~1000 chance with the current select.def.

There is also still an exception where if an authors name or "cluster" is in the catagory line it does not remove the other segments in the "vs line".  i also get the feeling that i may have made a spelling mistake.